Adjust the process for AiThread deleting
