+++ app-modules/ai/src/Events/AiThreadForceDeleting.php
@@ -4,12 +4,17 @@ namespace AdvisingApp\Ai\Events;
 
 use AdvisingApp\Ai\Models\AiThread;
 use Illuminate\Foundation\Events\Dispatchable;
+use AdvisingApp\Ai\Listeners\DeleteExternalAiThread;
+use AdvisingApp\Ai\Listeners\DeleteAiThreadVectorStores;
 
 class AiThreadForceDeleting
 {
     use Dispatchable;
 
-    public const LISTENERS = [];
+    public const LISTENERS = [
+        DeleteAiThreadVectorStores::class,
+        DeleteExternalAiThread::class,
+    ];
 
     public function __construct(public AiThread $aiThread) {}
 }
